UnitOfWork pattern added backend, modal confirm fe

+++ frontend/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -16,7 +17,7 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading = false;
 
-  constructor(private MessageService:MessageService) { }
+  constructor(private MessageService:MessageService, private confirmSErvice: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -37,8 +38,12 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id: number) {
-    this.MessageService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+    this.confirmSErvice.confirm('Confirm delete message', 'This can not be undone').subscribe(result => {
+      if(result) {
+        this.MessageService.deleteMessage(id).subscribe(() => {
+          this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+        })
+      }
     })
   }
